+++ tools/traffic_vault_util.go
@@ -37,14 +37,14 @@ var vault_user string
 var vault_pass string
 var vault_action string
 var dry_run bool
-var skip_tls bool
+var insecure bool
 var server_name string
 
-func connectToRiak(vault_ip string, vault_port uint, skip_tls bool, server_name string) *riak.Cluster {
+func connectToRiak(vault_ip string, vault_port uint, insecure bool, server_name string) *riak.Cluster {
 
 	tlsConfig := tls.Config{
 		ServerName:         server_name,
-		InsecureSkipVerify: skip_tls,
+		InsecureSkipVerify: insecure,
 	}
 
 	authOptions := riak.AuthOptions{
@@ -220,7 +220,7 @@ func init() {
 	flag.StringVar(&vault_pass, "vault_password", "", "Riak Password")
 	flag.StringVar(&vault_action, "vault_action", "", "Action: list_buckets|list_keys|list_values|convert_ssl_to_xmlid")
 	flag.BoolVar(&dry_run, "dry_run", false, "Do not perform writes")
-	flag.BoolVar(&skip_tls, "skip_tls", false, "Disable TLS certificate checks when connecting to cluster. Defaults to false")
+	flag.BoolVar(&insecure, "skip_tls", false, "Disable TLS certificate checks when connecting to cluster. Defaults to false")
 	flag.StringVar(&server_name, "server_name", "", "Name of server. Required and must match TLS configuration if 'skip_tls' is false.")
 }
 
@@ -236,11 +236,11 @@ func main() {
 		log.Fatal("Must provide Traffic Vault IP or host")
 	}
 
-	if !skip_tls && len(server_name) == 0 {
+	if !insecure && len(server_name) == 0 {
 		log.Fatal("Must provide valid Server Name if requiring TLS. To skip TLS verification, utilize 'skip_tls' flag.")
 	}
 
-	cluster := connectToRiak(vault_ip, vault_port, skip_tls, server_name)
+	cluster := connectToRiak(vault_ip, vault_port, insecure, server_name)
 	defer func() {
 		if err := cluster.Stop(); err != nil {
 			log.Fatal(err.Error())